Refactor pubgrub integration
