+++ mqpkg/src/resolver/types/requirement.rs
@@ -2,182 +2,13 @@
 // 2.0, and the BSD License. See the LICENSE file in the root of this repository
 // for complete details.
 
-use std::cmp::Ordering;
-use std::collections::HashMap;
 use std::fmt;
 
-use ::pubgrub::version::Version as PubGrubVersion;
-use ::pubgrub::version_set::VersionSet as PubGrubVersionSet;
-use dyn_clone::DynClone;
-use semver::{Prerelease, Version as SemVer, VersionReq};
+use pubgrub::version_set::VersionSet as BaseVersionSet;
+use semver::{Prerelease, VersionReq};
 
-use crate::resolver::pubgrub;
-use crate::resolver::Name;
-use crate::types::{RequestedPackages, Source, WithSource};
-
-pub(super) use super::pubgrub::VersionSet;
-
-pub(crate) trait Dependencies: fmt::Debug + DynClone {
-    fn get(&self) -> HashMap<Name, Requirement>;
-}
-
-dyn_clone::clone_trait_object!(Dependencies);
-
-#[derive(Debug, Clone)]
-struct StaticDependencies {
-    dependencies: HashMap<Name, Requirement>,
-}
-
-impl StaticDependencies {
-    fn new<N: Into<Name>>(dependencies: HashMap<N, Requirement>) -> StaticDependencies {
-        StaticDependencies {
-            dependencies: dependencies
-                .into_iter()
-                .map(|(p, r)| (p.into(), r))
-                .collect(),
-        }
-    }
-}
-
-impl Dependencies for StaticDependencies {
-    fn get(&self) -> HashMap<Name, Requirement> {
-        self.dependencies.clone()
-    }
-}
-
-pub(crate) trait WithDependencies {
-    fn dependencies(&self) -> &dyn Dependencies;
-}
-
-#[derive(Debug, Clone)]
-struct InternalSource(u64);
-
-impl InternalSource {
-    fn new(id: u64) -> InternalSource {
-        InternalSource(id)
-    }
-}
-
-impl fmt::Display for InternalSource {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Internal")
-    }
-}
-
-impl Source for InternalSource {
-    fn id(&self) -> u64 {
-        0
-    }
-
-    fn discriminator(&self) -> u64 {
-        self.0
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Version {
-    version: SemVer,
-    source_id: u64,
-    source_discriminator: u64,
-}
-
-impl Version {
-    fn new(major: u64, minor: u64, patch: u64) -> Version {
-        Version {
-            version: SemVer::new(major, minor, patch),
-            source_id: 0,
-            source_discriminator: 0,
-        }
-    }
-
-    fn candidate(major: u64, minor: u64, patch: u64) -> Version {
-        Version::new(major, minor, patch).with_source_id(u64::MAX)
-    }
-
-    fn pre<S: AsRef<str>>(mut self, pre: S) -> Version {
-        self.version.pre = Prerelease::new(pre.as_ref()).unwrap();
-        self
-    }
-
-    fn with_source_id(mut self, source_id: u64) -> Version {
-        self.source_id = source_id;
-        self
-    }
-
-    fn with_source_discriminator(mut self, source_discriminator: u64) -> Version {
-        self.source_discriminator = source_discriminator;
-        self
-    }
-}
-
-impl fmt::Display for Version {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.version)
-    }
-}
-
-impl PartialEq for Version {
-    fn eq(&self, other: &Self) -> bool {
-        (&self.version, self.source_id, self.source_discriminator)
-            == (&other.version, other.source_id, other.source_discriminator)
-    }
-}
-impl Eq for Version {}
-
-impl Ord for Version {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match self.version.cmp(&other.version) {
-            Ordering::Equal => (self.source_id, self.source_discriminator)
-                .cmp(&(other.source_id, other.source_discriminator))
-                .reverse(),
-            Ordering::Greater => Ordering::Greater,
-            Ordering::Less => Ordering::Less,
-        }
-    }
-}
-
-impl PartialOrd for Version {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl pubgrub::Version for Version {
-    fn is_prerelease(&self) -> bool {
-        !self.version.pre.is_empty()
-    }
-}
-
-impl PubGrubVersion for Version {
-    fn lowest() -> Version {
-        Version::candidate(0, 0, 0)
-    }
-
-    fn bump(&self) -> Version {
-        Version::new(
-            self.version.major,
-            self.version.minor,
-            self.version.patch + 1,
-        )
-        .with_source_id(self.source_id)
-    }
-}
-
-impl From<&SemVer> for Version {
-    fn from(version: &SemVer) -> Version {
-        Version {
-            version: version.clone(),
-            source_id: 0,
-            source_discriminator: 0,
-        }
-    }
-}
-
-impl From<&Version> for SemVer {
-    fn from(version: &Version) -> SemVer {
-        version.version.clone()
-    }
-}
+use crate::resolver::pubgrub::{Candidate, VersionSet};
+use crate::resolver::types::version::Version;
 
 #[derive(Debug, Clone)]
 pub(crate) struct Requirement(VersionReq);
@@ -197,94 +28,6 @@ impl fmt::Display for Requirement {
 impl From<VersionReq> for Requirement {
     fn from(req: VersionReq) -> Requirement {
         Requirement::new(req)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Candidate {
-    version: Version,
-    source: Box<dyn Source>,
-    dependencies: Box<dyn Dependencies + Sync + Send>,
-}
-
-impl Candidate {
-    pub(crate) fn new<V: Into<Version>>(
-        version: V,
-        source: Box<dyn Source>,
-        dependencies: Box<dyn Dependencies + Sync + Send>,
-    ) -> Candidate {
-        Candidate {
-            version: version
-                .into()
-                .with_source_id(source.id())
-                .with_source_discriminator(source.discriminator()),
-            source,
-            dependencies,
-        }
-    }
-
-    pub(super) fn root(reqs: RequestedPackages) -> Candidate {
-        Candidate {
-            version: Version::candidate(0, 0, 0),
-            source: Box::new(InternalSource::new(0)),
-            dependencies: Box::new(StaticDependencies::new(
-                reqs.iter()
-                    .map(|(k, v)| (k.clone(), Requirement(v.clone())))
-                    .collect(),
-            )),
-        }
-    }
-}
-
-impl WithDependencies for Candidate {
-    fn dependencies(&self) -> &dyn Dependencies {
-        &*self.dependencies
-    }
-}
-
-impl WithSource for Candidate {
-    fn source(&self) -> &Box<dyn Source> {
-        &self.source
-    }
-}
-
-impl fmt::Display for Candidate {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.version)
-    }
-}
-
-impl PartialEq for Candidate {
-    // It's important that this is just dispatched to Version, otherwise the
-    // internal Range will get confused.
-    fn eq(&self, other: &Candidate) -> bool {
-        self.version == other.version
-    }
-}
-
-impl Eq for Candidate {}
-
-impl Ord for Candidate {
-    // It's important that this is just dispatched to Version, otherwise the
-    // internal Range will get confused.
-    fn cmp(&self, other: &Candidate) -> Ordering {
-        self.version.cmp(&other.version)
-    }
-}
-
-impl PartialOrd for Candidate {
-    // It's important that this is just dispatched to Version, otherwise the
-    // internal Range will get confused.
-    fn partial_cmp(&self, other: &Candidate) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl pubgrub::Candidate for Candidate {
-    type V = Version;
-
-    fn version(&self) -> &Version {
-        &self.version
     }
 }
 
@@ -321,12 +64,6 @@ impl From<&Requirement> for VersionSet<Candidate> {
             vs = vs.with_normal(&convert_normal(comp));
             vs = vs.with_pre(&convert_prerelease(comp));
         }
-
-        // Since we're going to use only pre when checking against
-        // a pre-release, we still need to compute the intersection
-        // of what we would normally allow, against our union of
-        // of explicitly mentioned pre-releases.
-        // pre = range.intersection(&pre);
 
         vs
     }
